@@ -14,7 +14,7 @@ import json
 import time
 import urllib.parse
 from functools import wraps
-from typing import Any, Callable, Dict, List, Optional, Type, TypeVar, Union
+from typing import Any, Callable, Optional, Type, TypeVar
 from urllib.parse import urljoin
 from uuid import uuid1
 
@@ -43,7 +43,7 @@ def ensure_string(parameter_name: str) -> Callable:
         @lru_cache()
         def get_assets(
             self: "Market",
-            assets: Optional[Union[str, List[str]]] = None,
+            assets: Optional[str | list[str]] = None,
             aclass: Optional[str] = None,
         ) -> dict:
             # If the function was called using
@@ -96,7 +96,7 @@ class KrakenErrorHandler:
         """
         return self.__kexceptions.get_exception(msg)
 
-    def check(self: KrakenErrorHandler, data: dict) -> Union[dict, Any]:
+    def check(self: KrakenErrorHandler, data: dict) -> dict | Any:
         """
         Check if the error message is a known KrakenError response and than
         raise a custom exception or return the data containing the "error".
@@ -151,7 +151,7 @@ class KrakenErrorHandler:
         :rtype: dict
         """
         if "batchStatus" in data:
-            batch_status: List[Dict[str, Any]] = data["batchStatus"]
+            batch_status: list[dict[str, Any]] = data["batchStatus"]
             for status in batch_status:
                 if "status" in status:
                     exception: Type[KrakenException] = self.__get_exception(
@@ -186,6 +186,7 @@ class KrakenBaseSpotAPI:
         key: str = "",
         secret: str = "",
         url: str = "",
+        *,
         sandbox: bool = False,
         use_custom_exceptions: bool = True,
     ):
@@ -208,12 +209,13 @@ class KrakenBaseSpotAPI:
         self: KrakenBaseSpotAPI,
         method: str,
         uri: str,
-        timeout: int = 10,
-        auth: bool = True,
         params: Optional[dict] = None,
+        timeout: int = 10,
+        *,
+        auth: bool = True,
         do_json: bool = False,
         return_raw: bool = False,
-    ) -> Union[Dict[str, Any], List[str], List[Dict[str, Any]], requests.Response]:
+    ) -> dict[str, Any] | list[str] | list[dict[str, Any]] | requests.Response:
         """
         Handles the requested requests, by sending the request, handling the
         response, and returning the message or in case of an error the
@@ -223,13 +225,13 @@ class KrakenBaseSpotAPI:
         :type method: str
         :param uri: The endpoint to send the message
         :type uri: str
+        :param params: The query or post parameter of the request (default:
+            ``None``)
+        :type params: dict, optional
         :param timeout: Timeout for the request (default: ``10``)
         :type timeout: int
         :param auth: If the requests needs authentication (default: ``True``)
         :type auth: bool
-        :param params: The query or post parameter of the request (default:
-            ``None``)
-        :type params: dict, optional
         :param do_json: If the ``params`` must be "jsonified" - in case of
             nested dict style
         :type do_json: bool
@@ -287,36 +289,36 @@ class KrakenBaseSpotAPI:
         url: str = f"{self.url}{uri}"
         if method in ("GET", "DELETE"):
             return self.__check_response_data(
-                self.__session.request(
+                response=self.__session.request(
                     method=method,
                     url=url,
                     headers=headers,
                     timeout=timeout,
                 ),
-                return_raw,
+                return_raw=return_raw,
             )
 
         if do_json:
             return self.__check_response_data(
-                self.__session.request(
+                response=self.__session.request(
                     method=method,
                     url=url,
                     headers=headers,
                     json=params,
                     timeout=timeout,
                 ),
-                return_raw,
+                return_raw=return_raw,
             )
 
         return self.__check_response_data(
-            self.__session.request(
+            response=self.__session.request(
                 method=method,
                 url=url,
                 headers=headers,
                 data=params,
                 timeout=timeout,
             ),
-            return_raw,
+            return_raw=return_raw,
         )
 
     def _get_kraken_signature(
@@ -350,8 +352,9 @@ class KrakenBaseSpotAPI:
     def __check_response_data(
         self: KrakenBaseSpotAPI,
         response: requests.Response,
+        *,
         return_raw: bool = False,
-    ) -> Union[dict, list, requests.Response]:
+    ) -> dict | list | requests.Response:
         """
         Checks the response, handles the error (if exists) and returns the response data.
 
@@ -369,7 +372,7 @@ class KrakenBaseSpotAPI:
             if return_raw:
                 return response
             try:
-                data: Union[dict, list] = response.json()
+                data: dict | list = response.json()
             except ValueError as exc:
                 raise ValueError(response.content) from exc
 
@@ -395,7 +398,7 @@ class KrakenBaseSpotAPI:
     def __exit__(
         self: KrakenBaseSpotAPI,
         *exc: object,
-        **kwargs: Dict[str, Any],
+        **kwargs: dict[str, Any],
     ) -> None:
         pass
 
@@ -426,6 +429,7 @@ class KrakenBaseFuturesAPI:
         key: str = "",
         secret: str = "",
         url: str = "",
+        *,
         sandbox: bool = False,
         use_custom_exceptions: bool = True,
     ):
@@ -450,12 +454,13 @@ class KrakenBaseFuturesAPI:
         self: KrakenBaseFuturesAPI,
         method: str,
         uri: str,
-        timeout: int = 10,
-        auth: bool = True,
-        post_params: Optional[dict] = None,
         query_params: Optional[dict] = None,
+        post_params: Optional[dict] = None,
+        timeout: int = 10,
+        *,
+        auth: bool = True,
         return_raw: bool = False,
-    ) -> Union[Dict[str, Any], List[Dict[str, Any]], List[str], requests.Response]:
+    ) -> dict[str, Any] | list[dict[str, Any]] | list[str] | requests.Response:
         """
         Handles the requested requests, by sending the request, handling the
         response, and returning the message or in case of an error the
@@ -465,16 +470,16 @@ class KrakenBaseFuturesAPI:
         :type method: str
         :param uri: The endpoint to send the message
         :type uri: str
+        :param query_params: The query parameter of the request (default:
+            ``None``)
+        :type query_params: dict, optional
+        :param post_params: The query parameter of the request (default:
+            ``None``)
+        :type post_params: dict, optional
         :param timeout: Timeout for the request (default: ``10``)
         :type timeout: int
         :param auth: If the request needs authentication (default: ``True``)
         :type auth: bool
-        :param post_params: The query parameter of the request (default:
-            ``None``)
-        :type post_params: dict, optional
-        :param query_params: The query parameter of the request (default:
-            ``None``)
-        :type query_params: dict, optional
         :param do_json: If the ``post_params`` must be "jsonified" - in case of
             nested dict style
         :type do_json: bool, optional
@@ -490,7 +495,7 @@ class KrakenBaseFuturesAPI:
         method = method.upper()
 
         post_string: str = ""
-        strl: List[str]
+        strl: list[str]
         if post_params is not None:
             strl = [f"{key}={post_params[key]}" for key in sorted(post_params)]
             post_string = "&".join(strl)
@@ -529,7 +534,7 @@ class KrakenBaseFuturesAPI:
 
         if method in ("GET", "DELETE"):
             return self.__check_response_data(
-                self.__session.request(
+                response=self.__session.request(
                     method=method,
                     url=f"{self.url}{uri}"
                     if query_string == ""
@@ -537,30 +542,30 @@ class KrakenBaseFuturesAPI:
                     headers=headers,
                     timeout=timeout,
                 ),
-                return_raw,
+                return_raw=return_raw,
             )
 
         if method == "PUT":
             return self.__check_response_data(
-                self.__session.request(
+                response=self.__session.request(
                     method=method,
                     url=f"{self.url}{uri}",
                     params=str.encode(post_string),
                     headers=headers,
                     timeout=timeout,
                 ),
-                return_raw,
+                return_raw=return_raw,
             )
 
         return self.__check_response_data(
-            self.__session.request(
+            response=self.__session.request(
                 method=method,
                 url=f"{self.url}{uri}?{post_string}",
                 data=str.encode(post_string),
                 headers=headers,
                 timeout=timeout,
             ),
-            return_raw,
+            return_raw=return_raw,
         )
 
     def _get_kraken_futures_signature(
@@ -598,8 +603,9 @@ class KrakenBaseFuturesAPI:
     def __check_response_data(
         self: KrakenBaseFuturesAPI,
         response: requests.Response,
+        *,
         return_raw: bool = False,
-    ) -> Union[dict, requests.Response]:
+    ) -> dict | requests.Response:
         """
         Checks the response, handles the error (if exists) and returns the
         response data.
@@ -639,7 +645,7 @@ class KrakenBaseFuturesAPI:
     def __enter__(self: Self) -> Self:
         return self
 
-    def __exit__(self, *exc: object, **kwargs: Dict[str, Any]) -> None:
+    def __exit__(self, *exc: object, **kwargs: dict[str, Any]) -> None:
         pass
 
 
