@@ -8,7 +8,7 @@
 from __future__ import annotations
 
 from functools import lru_cache
-from typing import List, Optional, TypeVar, Union
+from typing import Optional, TypeVar
 
 from kraken.base_api import KrakenBaseFuturesAPI, defined
 
@@ -54,6 +54,7 @@ class Market(KrakenBaseFuturesAPI):
         key: str = "",
         secret: str = "",
         url: str = "",
+        *,
         sandbox: bool = False,
     ) -> None:
         super().__init__(key=key, secret=secret, url=url, sandbox=sandbox)
@@ -112,10 +113,10 @@ class Market(KrakenBaseFuturesAPI):
                 'more_candles': True
             }
         """
-        ttypes: tuple = ("spot", "mark", "trade")
+        tick_types: tuple = ("spot", "mark", "trade")
         resolutions: tuple = ("1m", "5m", "15m", "30m", "1h", "4h", "12h", "1d", "1w")
-        if tick_type not in ttypes:
-            raise ValueError(f"tick_type must be in {ttypes}")
+        if tick_type not in tick_types:
+            raise ValueError(f"tick_type must be in {tick_types}")
         if resolution is not None and resolution not in resolutions:
             raise ValueError(f"resolution must be in {resolutions}")
 
@@ -132,7 +133,7 @@ class Market(KrakenBaseFuturesAPI):
         )
 
     @lru_cache()
-    def get_tick_types(self: Market) -> List[str]:
+    def get_tick_types(self: Market) -> list[str]:
         """
         Retrieve the available tick types that can be used for example to access
         the :func:`kraken.futures.Market.get_ohlc` endpoint.
@@ -142,7 +143,7 @@ class Market(KrakenBaseFuturesAPI):
         This function uses caching. Run ``get_tick_types.cache_clear()`` to clear.
 
         :return: List of available tick types
-        :rtype: List[str]
+        :rtype: list[str]
 
         .. code-block:: python
             :linenos:
@@ -155,7 +156,7 @@ class Market(KrakenBaseFuturesAPI):
         return self._request(method="GET", uri="/api/charts/v1/", auth=False)  # type: ignore[return-value]
 
     @lru_cache()
-    def get_tradeable_products(self: Market, tick_type: str) -> List[str]:
+    def get_tradeable_products(self: Market, tick_type: str) -> list[str]:
         """
         Retrieve a list containing the tradeable assets on the futures market.
 
@@ -166,7 +167,7 @@ class Market(KrakenBaseFuturesAPI):
         :param tick_type: The kind of data, based on ``mark``, ``spot``, or ``trade``
         :type tick_type: str
         :return: List of tradeable assets
-        :type: List[str]
+        :type: list[str]
 
         .. code-block:: python
             :linenos:
@@ -183,7 +184,7 @@ class Market(KrakenBaseFuturesAPI):
         )
 
     @lru_cache()
-    def get_resolutions(self: Market, tick_type: str, tradeable: str) -> List[str]:
+    def get_resolutions(self: Market, tick_type: str, tradeable: str) -> list[str]:
         """
         Retrieve the list of available resolutions for a specific asset.
 
@@ -196,7 +197,7 @@ class Market(KrakenBaseFuturesAPI):
         :param tick_type: The asset of interest
         :type tick_type: str
         :return: List of resolutions
-        :type: List[str]
+        :type: list[str]
 
         .. code-block:: python
             :linenos:
@@ -640,7 +641,7 @@ class Market(KrakenBaseFuturesAPI):
     def set_leverage_preference(
         self: Market,
         symbol: str,
-        maxLeverage: Optional[Union[str, int, float]] = None,
+        maxLeverage: Optional[str | float] = None,
     ) -> dict:
         """
         Set a new leverage preference for a specific futures contract.
@@ -652,7 +653,7 @@ class Market(KrakenBaseFuturesAPI):
         :param symbol: The symbol to set the preference
         :type symbol: str, optional
         :param maxLeverage: The maximum allowed leverage for a futures contract
-        :type maxLeverage: str | int | float, optional
+        :type maxLeverage: str | float, optional
         :return: Information about the success or fail
         :rtype: dict
 
