@@ -13,7 +13,7 @@ from asyncio import sleep as asyncio_sleep
 from binascii import crc32
 from collections import OrderedDict
 from inspect import iscoroutinefunction
-from typing import Callable, Dict, List, Optional, Union
+from typing import Callable, Optional
 
 from kraken.spot import Market
 from kraken.spot.websocket_v2 import KrakenSpotWSClientV2
@@ -116,7 +116,7 @@ class OrderbookClient:
         callback: Optional[Callable] = None,
     ) -> None:
         super().__init__()
-        self.__book: Dict[str, dict] = {}
+        self.__book: dict[str, dict] = {}
         self.__depth: int = depth
         self.__callback: Optional[Callable] = callback
 
@@ -125,7 +125,7 @@ class OrderbookClient:
             callback=self.on_message,
         )
 
-    async def on_message(self: OrderbookClient, message: Union[list, dict]) -> None:
+    async def on_message(self: OrderbookClient, message: list | dict) -> None:
         """
         The on_message function is implemented within the KrakenSpotWSClient
         class and used as callback to receive all messages sent by the
@@ -235,7 +235,7 @@ class OrderbookClient:
         else:
             print(message)  # noqa: T201
 
-    async def add_book(self: OrderbookClient, pairs: List[str]) -> None:
+    async def add_book(self: OrderbookClient, pairs: list[str]) -> None:
         """
         Add an orderbook to this client. The feed will be subscribed
         and updates will be published to the :func:`on_book_update` function.
@@ -249,7 +249,7 @@ class OrderbookClient:
             params={"channel": "book", "depth": self.__depth, "symbol": pairs},
         )
 
-    async def remove_book(self: OrderbookClient, pairs: List[str]) -> None:
+    async def remove_book(self: OrderbookClient, pairs: list[str]) -> None:
         """
         Unsubscribe from a subscribed orderbook.
 
@@ -314,7 +314,7 @@ class OrderbookClient:
 
     def __update_book(
         self: OrderbookClient,
-        orders: List[dict],
+        orders: list[dict],
         side: str,
         symbol: str,
         timestamp: Optional[str] = None,
