@@ -11,7 +11,7 @@ from __future__ import annotations
 from decimal import Decimal
 from functools import lru_cache
 from math import floor
-from typing import List, Optional, TypeVar, Union
+from typing import Optional, TypeVar
 
 from kraken.base_api import KrakenBaseSpotAPI, defined, ensure_string
 from kraken.spot.market import Market
@@ -69,25 +69,26 @@ class Trade(KrakenBaseSpotAPI):
         ordertype: str,
         side: str,
         pair: str,
-        volume: Union[str, float],
-        price: Optional[Union[str, float]] = None,
-        price2: Optional[Union[str, float]] = None,
-        truncate: bool = False,
+        volume: str | float,
+        price: Optional[str | float] = None,
+        price2: Optional[str | float] = None,
         trigger: Optional[str] = None,
         leverage: Optional[str] = None,
-        reduce_only: Optional[bool] = False,
         stptype: Optional[str] = "cancel-newest",
-        oflags: Optional[Union[str, List[str]]] = None,
+        oflags: Optional[str | list[str]] = None,
         timeinforce: Optional[str] = None,
         displayvol: Optional[str] = None,
         starttm: Optional[str] = "0",
         expiretm: Optional[str] = None,
         close_ordertype: Optional[str] = None,
-        close_price: Optional[Union[str, float]] = None,
-        close_price2: Optional[Union[str, float]] = None,
+        close_price: Optional[str | float] = None,
+        close_price2: Optional[str | float] = None,
         deadline: Optional[str] = None,
-        validate: bool = False,
         userref: Optional[int] = None,
+        *,
+        truncate: bool = False,
+        reduce_only: Optional[bool] = False,
+        validate: bool = False,
     ) -> dict:
         """
         Create a new order and place it on the market.
@@ -116,18 +117,12 @@ class Trade(KrakenBaseSpotAPI):
                 * Prefixed by # is the same as ``+`` and ``-`` but the sign is set automatically
                 * The percentage sign ``%`` can be used to define relative changes.
         :type price2: str | float, optional
-        :param truncate: If enabled: round the ``price`` and ``volume`` to Kraken's
-            maximum allowed decimal places. See https://support.kraken.com/hc/en-us/articles/4521313131540
-            fore more information about decimals.
-        :type truncate: bool, optional
         :param trigger: What triggers the position of ``stop-loss``, ``stop-loss-limit``, ``take-profit``, and
             ``take-profit-limit`` orders. Will also be used for associated conditional close orders.
             Kraken will use ``last`` if nothing is specified.
         :type trigger: str, optional
         :param leverage: The leverage
         :type leverage: str | float, optional
-        :param reduce_only: Reduce existing orders (default: ``False``)
-        :type reduce_only: bool, optional
         :param stptype: Define what cancels the order, one of ``cancel-newest``,
             ``cancel-oldest``, ``cancel-both`` (default: ``cancel-newest``)
         :type stptype: str, optional
@@ -155,6 +150,12 @@ class Trade(KrakenBaseSpotAPI):
         :param deadline: RFC3339 timestamp + {0..60} seconds that defines when the matching
             engine should reject the order.
         :type deadline: str, optional
+        :param truncate: If enabled: round the ``price`` and ``volume`` to Kraken's
+            maximum allowed decimal places. See https://support.kraken.com/hc/en-us/articles/4521313131540
+            fore more information about decimals.
+        :type truncate: bool, optional
+        :param reduce_only: Reduce existing orders (default: ``False``)
+        :type reduce_only: bool, optional
         :param validate: Validate the order without placing on the market (default: ``False``)
         :type validate: bool, optional
         :param userref: User reference id for example to group orders
@@ -364,9 +365,10 @@ class Trade(KrakenBaseSpotAPI):
 
     def create_order_batch(
         self: Trade,
-        orders: List[dict],
+        orders: list[dict],
         pair: str,
         deadline: Optional[str] = None,
+        *,
         validate: bool = False,
     ) -> dict:
         """
@@ -445,15 +447,16 @@ class Trade(KrakenBaseSpotAPI):
         self: Trade,
         txid: str,
         pair: str,
-        volume: Optional[Union[str, int, float]] = None,
-        price: Optional[Union[str, int, float]] = None,
-        price2: Optional[Union[str, int, float]] = None,
-        truncate: bool = False,
+        volume: Optional[str | float] = None,
+        price: Optional[str | float] = None,
+        price2: Optional[str | float] = None,
         oflags: Optional[str] = None,
         deadline: Optional[str] = None,
         cancel_response: Optional[bool] = None,
-        validate: bool = False,
         userref: Optional[int] = None,
+        *,
+        truncate: bool = False,
+        validate: bool = False,
     ) -> dict:
         """
         Edit an open order.
@@ -468,15 +471,11 @@ class Trade(KrakenBaseSpotAPI):
         :param pair: The asset pair of the order
         :type pair: str
         :param volume: Set a new volume
-        :type volume: str | int | float, optional
+        :type volume: str | float, optional
         :param price: Set a new price
-        :type price: str | int | float, optional
+        :type price: str | float, optional
         :param price2: Set a new second price
-        :type price2: str | int | float, optional
-        :param truncate: If enabled: round the ``price`` and ``volume`` to Kraken's
-            maximum allowed decimal places. See https://support.kraken.com/hc/en-us/articles/4521313131540
-            fore more information about decimals.
-        :type truncate: bool, optional
+        :type price2: str | float, optional
         :param oflags: Order flags like ``post``, ``fcib``, ``fciq``, ``nomp``,
             ``viqc`` (see the referenced Kraken documentation for more information)
         :type oflags: str | List[str], optional
@@ -484,6 +483,10 @@ class Trade(KrakenBaseSpotAPI):
         :type deadline: string
         :param cancel_response: See the referenced Kraken documentation for more information
         :type cancel_response: bool, optional
+        :param truncate: If enabled: round the ``price`` and ``volume`` to Kraken's
+            maximum allowed decimal places. See https://support.kraken.com/hc/en-us/articles/4521313131540
+            fore more information about decimals.
+        :type truncate: bool, optional
         :param validate: Validate the order without placing on the market (default: ``False``)
         :type validate: bool, optional
         :param userref: User reference id for example to group orders
@@ -632,7 +635,7 @@ class Trade(KrakenBaseSpotAPI):
             params={"timeout": timeout},
         )
 
-    def cancel_order_batch(self: Trade, orders: List[Union[str, int]]) -> dict:
+    def cancel_order_batch(self: Trade, orders: list[str | int]) -> dict:
         """
         Cancel a a list of orders by ``txid`` or ``userref``
 
@@ -642,7 +645,7 @@ class Trade(KrakenBaseSpotAPI):
         - https://docs.kraken.com/rest/#operation/cancelOrderBatch
 
         :param orders: List of orders to cancel
-        :type orders: List[str | int]
+        :type orders: list[str | int]
         :return: Success or failure - Number of closed orders
         :rtype: dict
 
@@ -667,7 +670,7 @@ class Trade(KrakenBaseSpotAPI):
     @lru_cache()
     def truncate(
         self: Trade,
-        amount: Union[Decimal, float, str],
+        amount: Decimal | float | str,
         amount_type: str,
         pair: str,
     ) -> str:
