@@ -14,7 +14,7 @@ from __future__ import annotations
 import asyncio
 import json
 from copy import deepcopy
-from typing import Any, Callable, List, Optional, Union
+from typing import Any, Callable, Optional
 
 from kraken.base_api import defined, ensure_string
 from kraken.exceptions import KrakenException
@@ -160,6 +160,7 @@ class KrakenSpotWSClient(KrakenSpotWSClientBase):
         key: str = "",
         secret: str = "",
         callback: Optional[Callable] = None,
+        *,
         no_public: bool = False,
         beta: bool = False,
     ):
@@ -175,6 +176,7 @@ class KrakenSpotWSClient(KrakenSpotWSClientBase):
     async def send_message(  # pylint: disable=arguments-differ
         self: KrakenSpotWSClient,
         message: dict,
+        *,
         private: bool = False,
         raw: bool = False,
     ) -> None:
@@ -215,7 +217,7 @@ class KrakenSpotWSClient(KrakenSpotWSClientBase):
     async def subscribe(  # pylint: disable=arguments-differ
         self: KrakenSpotWSClient,
         subscription: dict,
-        pair: Optional[List[str]] = None,
+        pair: Optional[list[str]] = None,
     ) -> None:
         """
         Subscribe to a channel
@@ -287,7 +289,7 @@ class KrakenSpotWSClient(KrakenSpotWSClientBase):
     async def unsubscribe(  # pylint: disable=arguments-differ
         self: KrakenSpotWSClient,
         subscription: dict,
-        pair: Optional[List[str]] = None,
+        pair: Optional[list[str]] = None,
     ) -> None:
         """
         Unsubscribe from a feed
@@ -357,7 +359,7 @@ class KrakenSpotWSClient(KrakenSpotWSClientBase):
             # await self.send_message(payload, private=False)
 
     @property
-    def public_channel_names(self: KrakenSpotWSClient) -> List[str]:
+    def public_channel_names(self: KrakenSpotWSClient) -> list[str]:
         """
         Returns the public subscription names
 
@@ -368,7 +370,7 @@ class KrakenSpotWSClient(KrakenSpotWSClientBase):
         return ["ticker", "spread", "book", "ohlc", "trade", "*"]
 
     @property
-    def private_channel_names(self: KrakenSpotWSClient) -> List[str]:
+    def private_channel_names(self: KrakenSpotWSClient) -> list[str]:
         """
         Returns the private subscription names
 
@@ -384,21 +386,22 @@ class KrakenSpotWSClient(KrakenSpotWSClientBase):
         ordertype: str,
         side: str,
         pair: str,
-        volume: Union[str, float],
-        price: Optional[Union[str, float]] = None,
-        price2: Optional[Union[str, float]] = None,
-        truncate: bool = False,
-        leverage: Optional[Union[str, float]] = None,
-        oflags: Optional[Union[str, List[str]]] = None,
-        starttm: Optional[Union[str, int]] = None,
-        expiretm: Optional[Union[str, int]] = None,
+        volume: str | float,
+        price: Optional[str | float] = None,
+        price2: Optional[str | float] = None,
+        leverage: Optional[str | float] = None,
+        oflags: Optional[str | list[str]] = None,
+        starttm: Optional[str | int] = None,
+        expiretm: Optional[str | int] = None,
         deadline: Optional[str] = None,
-        userref: Optional[Union[str, int]] = None,
-        validate: bool = False,
+        userref: Optional[str | int] = None,
         close_ordertype: Optional[str] = None,
-        close_price: Optional[Union[str, int, float]] = None,
-        close_price2: Optional[Union[str, int, float]] = None,
-        timeinforce: Optional[Union[str, int]] = None,
+        close_price: Optional[str | float] = None,
+        close_price2: Optional[str | float] = None,
+        timeinforce: Optional[str | int] = None,
+        *,
+        truncate: bool = False,
+        validate: bool = False,
     ) -> None:
         """
         Create an order and submit it.
@@ -427,11 +430,6 @@ class KrakenSpotWSClient(KrakenSpotWSClientBase):
             ``take-profit-limit`` orders (see the referenced Kraken
             documentation for more information)
         :type price2: str | float, optional
-        :param truncate: If enabled: round the ``price`` and ``volume`` to
-            Kraken's maximum allowed decimal places. See
-            https://support.kraken.com/hc/en-us/articles/4521313131540
-            fore more information about decimals.
-        :type truncate: bool, optional
         :param leverage: The leverage
         :type leverage: str | float, optional
         :param oflags: Order flags like ``post``, ``fcib``, ``fciq``, ``nomp``,
@@ -449,9 +447,6 @@ class KrakenSpotWSClient(KrakenSpotWSClientBase):
         :type deadline: str
         :param userref: User reference id for example to group orders
         :type userref: int
-        :param validate: Validate the order without placing on the market
-            (default: ``False``)
-        :type validate: bool, optional
         :param close_ordertype:  Conditional close order type, one of:
             ``limit``, ``stop-loss``, ``take-profit``, ``stop-loss-limit``,
             ``take-profit-limit``
@@ -464,6 +459,14 @@ class KrakenSpotWSClient(KrakenSpotWSClientBase):
             ``GTC``, ``IOC``, ``GTD`` (see the referenced Kraken documentation
             for more information)
         :type timeinforce: str, optional
+        :param truncate: If enabled: round the ``price`` and ``volume`` to
+            Kraken's maximum allowed decimal places. See
+            https://support.kraken.com/hc/en-us/articles/4521313131540
+            fore more information about decimals.
+        :type truncate: bool, optional
+        :param validate: Validate the order without placing on the market
+            (default: ``False``)
+        :type validate: bool, optional
         :raises KrakenAuthenticationError: If the websocket is not connected or
             the connection is not authenticated
         :raises ValueError: If input is not correct
@@ -548,14 +551,15 @@ class KrakenSpotWSClient(KrakenSpotWSClientBase):
     async def edit_order(  # noqa: PLR0913
         self: KrakenSpotWSClient,
         orderid: str,
-        reqid: Optional[Union[str, int]] = None,
+        reqid: Optional[str | int] = None,
         pair: Optional[str] = None,
-        price: Optional[Union[str, int, float]] = None,
-        price2: Optional[Union[str, int, float]] = None,
+        price: Optional[str | float] = None,
+        price2: Optional[str | float] = None,
+        volume: Optional[str | float] = None,
+        oflags: Optional[str | list[str]] = None,
+        newuserref: Optional[str | int] = None,
+        *,
         truncate: bool = False,
-        volume: Optional[Union[str, int, float]] = None,
-        oflags: Optional[Union[str, List[str]]] = None,
-        newuserref: Optional[Union[str, int]] = None,
         validate: bool = False,
     ) -> None:
         """
@@ -576,17 +580,17 @@ class KrakenSpotWSClient(KrakenSpotWSClientBase):
         :type price: str | int | float, optional
         :param price2: Set a new second price
         :type price2: str | int | float, optional
-        :param truncate: If enabled: round the ``price`` and ``volume`` to Kraken's
-            maximum allowed decimal places. See
-            https://support.kraken.com/hc/en-us/articles/4521313131540 fore more
-            information about decimals.
-        :type truncate: bool, optional
         :param volume: Set a new volume
         :type volume: str | int | float, optional
         :param oflags: Set new oflags (overwrite old ones)
         :type oflags: str | list[str], optional
         :param newuserref: Set a new user reference id
         :type newuserref: str | int, optional
+        :param truncate: If enabled: round the ``price`` and ``volume`` to Kraken's
+            maximum allowed decimal places. See
+            https://support.kraken.com/hc/en-us/articles/4521313131540 fore more
+            information about decimals.
+        :type truncate: bool, optional
         :param validate: Validate the input without applying the changes
             (default: ``False``)
         :type validate: bool, optional
@@ -644,7 +648,7 @@ class KrakenSpotWSClient(KrakenSpotWSClientBase):
 
         await self.send_message(message=payload, private=True)
 
-    async def cancel_order(self: KrakenSpotWSClient, txid: List[str]) -> None:
+    async def cancel_order(self: KrakenSpotWSClient, txid: list[str]) -> None:
         """
         Cancel a specific order or a list of orders.
 
