@@ -2,10 +2,10 @@
 # -*- coding: utf-8 -*-
 # Copyright (C) 2023 Benjamin Thomas Schwertfeger
 # GitHub: https://github.com/btschwertfeger
+# pylint: disable=unused-import
 
 """Module that provides the Spot REST clients and utility functions."""
 
-# pylint: disable=unused-import
 from kraken.spot.funding import Funding
 from kraken.spot.market import Market
 from kraken.spot.orderbook import OrderbookClient
